+++ riscv-macros/CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## v0.4.1 - Unreleased
+## Unreleased
+
+## v0.4.1 - 2026-05-29
 
 ### Added